requirements.txt
sentence-transformers
uvicorn
ctransformers
langchain
python-box
streamlit
google-generativeai
python-dotenv
